更新bug 10-2.py 和10-3.py传反了
